+++ src/prelude.rs
@@ -2,7 +2,9 @@
 
 pub use crate::animatable::Animatable;
 pub use crate::keyframes::Keyframes;
-pub use crate::property::{PropertyValue, UiProperty};
+pub use crate::property::{
+    PropertyValue, RectangleValue, SizeValue, TransformValue, UiProperty, Vector2Value,
+};
 pub use crate::runtime::AnimationRuntime;
 pub use crate::timeline::Timeline;
 pub use crate::timing::Timing;